Fixade så att alla saker har doubles som koordinater istället för ints. Fixade också en timer som används när spelaren rör på sig.

+++ src/model/ThingModelInterface.java
@@ -6,8 +6,8 @@ package model;
  * a representation in the view.
  */
 public interface ThingModelInterface {
-	public int getX();
-	public int getY();
+	public double getX();
+	public double getY();
 	public void update();
 	public ThingState getState();
 	public void changeState(ThingState state);
